+++ src/bucket/link.rs
@@ -246,7 +246,6 @@ mod tests {
     }
 
     #[rstest]
-    #[cfg(feature = "test-api-119")]
     #[tokio::test]
     async fn test_link_creation_with_explicit_record_identity(#[future] bucket: Bucket) {
         let bucket: Bucket = bucket.await;